+++ src-tauri/src/lib.rs
@@ -1,0 +1,31 @@
+use itertools::Itertools;
+
+#[cfg_attr(mobile, tauri::mobile_entry_point)]
+pub fn run() {
+    tauri::Builder::default()
+        .plugin(tauri_plugin_opener::init())
+        .setup(|app| {
+            if cfg!(debug_assertions) {
+                app.handle().plugin(
+                    tauri_plugin_log::Builder::default()
+                        .level(log::LevelFilter::Info)
+                        .build(),
+                )?;
+            }
+            Ok(())
+        })
+        .invoke_handler(tauri::generate_handler![list_fonts])
+        .run(tauri::generate_context!())
+        .expect("error while running tauri application");
+}
+
+#[tauri::command]
+fn list_fonts() -> Vec<String> {
+    font_enumeration::Collection::new()
+        .unwrap()
+        .all()
+        .into_iter()
+        .map(|f| f.family_name.clone())
+        .unique()
+        .collect()
+}