+++ upload-itch.ps1
@@ -1,11 +1,15 @@
 # Note: itch.io butler needs to be installed for this script to work.
 # It requires authentication so the script is only usable by admins ;)
 
-# Build
-npm run electron:build
+# Build Windows
+npm ci
+npx tauri build
+
+# Build Linux
+wsl -- bash -ic "npm ci && npx tauri build"
 
 # Upload to itch
-butler push electron_build\calligro-win32-x64 voycawojka/calligro:win-64 --userversion-file version.txt
-butler push electron_build\calligro-linux-x64 voycawojka/calligro:linux --userversion-file version.txt
+butler push src-tauri\target\release\calligro.exe voycawojka/calligro:win-64 --userversion-file version.txt
+butler push src-tauri\target\release\calligro voycawojka/calligro:linux --userversion-file version.txt
 
 Write-Host "Upload done"